Полная переработка передачи данных
