+++ app.py
@@ -1,6 +1,7 @@
 from flask import Flask, render_template, url_for, request
-from get_weather_variables import today, hours, week
-from get_weather_json import city
+from datetime import date
+from get_weather_variables import get_all_weather_data
+from classes import City
 
 app = Flask(__name__)
 
@@ -8,15 +9,17 @@ app = Flask(__name__)
 @app.route("/", methods=["GET", "POST"])
 def index():
     if request.method == "POST":
-        pass
-        # latitude = request.form.get("latitude")
-        # longitude = request.form.get("longitude")
-    return render_template('index.html', weather_url=url_for("weather"))
+        city = City(request.form.get("latitude"), request.form.get(
+            "longitude"), request.form.get("name"))
+        cur_date = date.today()
+        data = get_all_weather_data(city, cur_date)
+        return render_template('weather.html', city=city.name, today=data["today"], hours=data["hours"], week=data["week"])
+    return render_template('index.html')
 
 
-@app.route("/weather")
-def weather():
-    return render_template('weather.html', city=city.name, today=today, hours=hours, week=week)
+# @app.route("/weather")
+# def weather():
+#     return render_template('weather.html', city=city.name, today=today, hours=hours, week=week)
 
 
 if __name__ == "__main__":
