+++ templates/index.html
@@ -6,6 +6,17 @@
     <title>Document</title>
   </head>
   <body>
-    <h1><a href="{{ weather_url }}">Link</a></h1>
+    <form action="{{ url_for('index') }}" method="post">
+      <label for="latitude">latitude:</label>
+      <input type="text" id="latitude" name="latitude" required /><br /><br />
+
+      <label for="longitude">longitude:</label>
+      <input type="text" id="longitude" name="longitude" required /><br /><br />
+
+      <label for="name">Name:</label>
+      <input type="text" id="name" name="name" required /><br /><br />
+
+      <input type="submit" value="Submit" />
+    </form>
   </body>
 </html>
